requirements.txt
fastapi
uvicorn
bcrypt
pymongo
python-dotenv
python-jose
motor
cloudinary
python-multipart